clarified documentations, corrected style errors.
