+++ plaso/parsers/ios_mobile_installation_logs.py
@@ -1,6 +1,8 @@
 # -*- coding: utf-8 -*-
 """Parser for iOS mobile installation logs files obtained from iOS device.
-Based on the research from Alex Brignoni."""
+Based on the research by Alex Brignoni.
+See https://dfir.pubpub.org/pub/e5xlbw88/release/2 for the paper.
+Also https://github.com/abrignoni/iOS-Mobile-Installation-Logs-Parser."""
 
 import pyparsing
 
@@ -16,11 +18,11 @@ from plaso.parsers import text_parser
 
 class MobileInstallationEventData(events.EventData):
   """Mobile Installation log event data
+
   Attributes:
-    number (str):
-    severity (str):
-    id (str):
-    originator (str): process that created the entry
+    process_identifier (str): process_identifier
+    severity (str): severity of the message
+    originating_call (str): call that created the entry
     body (str): body of the event line
   """
 
@@ -29,31 +31,31 @@ class MobileInstallationEventData(events.EventData):
   def __init__(self):
     """Initializes event data."""
     super(MobileInstallationEventData, self).__init__(data_type=self.DATA_TYPE)
-    self.number = None
+    self.process_identifier = None
     self.severity = None
-    self.id = None
-    self.originator = None
+    self.originating_call = None
     self.body = None
 
 
 class MobileInstallationParser(text_parser.PyparsingMultiLineTextParser):
   """Parser for iOS Mobile Installation logs."""
 
-  NAME = 'mobile_installation.log'
+  NAME = 'ios_mobile_installation.log'
   DATA_FORMAT = 'iOS mobile installation logs'
 
-  MONTHS = {'Jan': 1,
-            'Feb': 2,
-            'Mar': 3,
-            'Apr': 4,
-            'May': 5,
-            'Jun': 6,
-            'Jul': 7,
-            'Aug': 8,
-            'Sep': 9,
-            'Oct': 10,
-            'Nov': 11,
-            'Dec': 12}
+  MONTHS = {
+      'Jan': 1,
+      'Feb': 2,
+      'Mar': 3,
+      'Apr': 4,
+      'May': 5,
+      'Jun': 6,
+      'Jul': 7,
+      'Aug': 8,
+      'Sep': 9,
+      'Oct': 10,
+      'Nov': 11,
+      'Dec': 12}
 
   ONE_OR_TWO_DIGITS = text_parser.PyparsingConstants.ONE_OR_TWO_DIGITS
   FOUR_DIGITS = text_parser.PyparsingConstants.FOUR_DIGITS
@@ -61,31 +63,29 @@ class MobileInstallationParser(text_parser.PyparsingMultiLineTextParser):
   TIME_ELEMENTS = text_parser.PyparsingConstants.TIME_ELEMENTS
 
   _TIMESTAMP = (THREE_LETTERS.suppress() + THREE_LETTERS('month') +
-                ONE_OR_TWO_DIGITS('day') + TIME_ELEMENTS + FOUR_DIGITS('year'))
+      ONE_OR_TWO_DIGITS('day') + TIME_ELEMENTS + FOUR_DIGITS('year'))
 
-  _ELEMENT_NUMBER = pyparsing.Suppress('[') + \
-                    pyparsing.Word(pyparsing.nums)('number') + \
-                    pyparsing.Suppress(']')
+  _ELEMENT_NUMBER = (pyparsing.Suppress('[') +
+      pyparsing.Word(pyparsing.nums)('process_identifier') +
+      pyparsing.Suppress(']'))
 
-  _ELEMENT_SEVERITY = pyparsing.Suppress('<') + \
-                      pyparsing.Word(pyparsing.alphanums)('severity') + \
-                      pyparsing.Suppress('>')
+  _ELEMENT_SEVERITY = (pyparsing.Suppress('<') +
+      pyparsing.Word(pyparsing.alphanums)('severity') + pyparsing.Suppress('>'))
 
-  _ELEMENT_ID = pyparsing.Suppress('(') + \
-                pyparsing.Word(pyparsing.alphanums)('id') + \
-                pyparsing.Suppress(')')
+  _ELEMENT_ID = (pyparsing.Suppress('(') +
+      pyparsing.Word(pyparsing.alphanums)('id') + pyparsing.Suppress(')'))
 
-  _ELEMENT_ORIGINATOR = (pyparsing.Suppress('+[') | pyparsing.Suppress('-[')) \
-      + pyparsing.SkipTo(pyparsing.Literal(']'))('originator') \
-      + pyparsing.Suppress(']: ')
+  _ELEMENT_ORIGINATOR = ((pyparsing.Suppress('+[') | pyparsing.Suppress('-[')) +
+      pyparsing.SkipTo(pyparsing.Literal(']'))('originating_call') +
+      pyparsing.Suppress(']: '))
 
   _BODY_END = pyparsing.StringEnd() | _TIMESTAMP
 
   _ELEMENT_BODY = pyparsing.SkipTo(_BODY_END)('body')
 
-  _LINE_GRAMMAR = _TIMESTAMP + _ELEMENT_NUMBER + _ELEMENT_SEVERITY + \
-                  _ELEMENT_ID + _ELEMENT_ORIGINATOR + \
-                  _ELEMENT_BODY + pyparsing.ZeroOrMore(pyparsing.lineEnd())
+  _LINE_GRAMMAR = (_TIMESTAMP + _ELEMENT_NUMBER + _ELEMENT_SEVERITY +
+      _ELEMENT_ID + _ELEMENT_ORIGINATOR + _ELEMENT_BODY +
+      pyparsing.ZeroOrMore(pyparsing.lineEnd()))
 
   LINE_STRUCTURES = [('log_entry', _LINE_GRAMMAR)]
 
@@ -107,7 +107,7 @@ class MobileInstallationParser(text_parser.PyparsingMultiLineTextParser):
 
     if key != 'log_entry':
       raise errors.ParseError(
-        'Unable to parse record, unknown structure: {0:s}'.format(key))
+          'Unable to parse record, unknown structure: {0:s}'.format(key))
 
     year = self._GetValueFromStructure(structure, 'year')
     month = self.MONTHS.get(self._GetValueFromStructure(structure, 'month'))
@@ -117,15 +117,16 @@ class MobileInstallationParser(text_parser.PyparsingMultiLineTextParser):
     seconds = self._GetValueFromStructure(structure, 'seconds')
 
     event_data = MobileInstallationEventData()
-    event_data.number = self._GetValueFromStructure(structure, 'number')
+    event_data.process_identifier = self._GetValueFromStructure(
+        structure, 'process_identifier')
     event_data.severity = self._GetValueFromStructure(structure, 'severity')
-    event_data.id = self._GetValueFromStructure(structure, 'id')
-    event_data.originator = self._GetValueFromStructure(structure, 'originator')
+    event_data.originating_call = self._GetValueFromStructure(
+        structure, 'originating_call')
     event_data.body = self._GetValueFromStructure(structure, 'body')
 
     try:
       date_time = dfdatetime_time_elements.TimeElements(time_elements_tuple=(
-        year, month, day, hours, minutes, seconds))
+          year, month, day, hours, minutes, seconds))
     except (TypeError, ValueError):
       parser_mediator.ProduceExtractionWarning('invalid date time value')
       return
